Return init cgroup path relative to the namespace mount.
If the cgroup path in /proc/1/cgroup contains .., then it is the
relative path to the root cgroup of init's cgroup namespace. Since
the relative part of the path is not useful, it is removed, leaving
only the path to init's cgroup. The return value is formed by joining
the configured cgroup mount path (usually /sys/fs/cgroup) and the
remaining part of the path to init's cgroup.

Signed-off-by: Urban Borštnik <urban.borstnik@id.ethz.ch>