Get the relative path to the cgroup namespace root.
In a cgroup namespace the paths in /proc/*/cgroup are relative
to the root of the cgroup namespace. The _get_relative_cg_path
function returns this relative path so that it can be removed
from the values read in /proc/*/cgroup.

This function reads the cgroup path for the given proccess PID in
/proc/PID/mountinfo.

Signed-off-by: Urban Borštnik <urban.borstnik@id.ethz.ch>